Add files via upload
For context, this MacBook Pro and Mac mini have developer mode forced on (not by me). They use ‘personas’ to be able to circumvent permission restrictions. The root kernel has fake certification certificates.

This log is interesting for a number of reasons:

- There are many non-standard operations being performed. I installed Bitdefender, which was a mistake as they modified it.

- It details triald (Trial), (TrialServer), TRIClient, TrialXP activities: com.apple.siri.embeddedspeech: (Trial) [com.apple.trial:_] Initializing TRIClient. Trial version: TrialXP-347.4.1

- It documents attempts to activate FaceTime messaging, despite it being disabled by settings in the Control Panel. It ends up succeeding and is still using FT to convey messages, diverting Lockdown by having added an unknown hidden contact in the contact address book.

- They use gzip to upload my Notes, despite backing up to iCloud (or anywhere) disabled.

- They use Siri embedded speech to remotely load changes (experiments) to nearby iOS devices: triald: (TrialServer) [com.apple.triald:server] TRIXPCNamespaceManagementRequestHandler 0x12981b0f0: begin statusOfDownloadForFactors:(
    "com.apple.siri.asr.assistant.en_CA"
) withNamespace:SIRI_UNDERSTANDING_ASR_ASSISTANT fstate:(null) qos:21 completion:

iPad example (this is documented in another log provided):
  "legacyInfo" : {
  "threadTriggered" : {
    "queue" : "com.facebook.react.ExceptionsManagerQueue"
  }
},
  "logWritingSignature" : "RedactedOnPurpose",
  "trialInfo" : {
  "rollouts" : [
    {
      "rolloutId" : "RedactedOnPurpose",
      "factorPackIds" : {
        "SIRI_TEXT_TO_SPEECH" : "RedactedOnPurpose"
      },
      "deploymentId" : RedactedOnPurpose
    },
    {
      "rolloutId" : "RedactedOnPurpose",
      "factorPackIds" : {
        "SIRI_FIND_MY_CONFIGURATION_FILES" : "RedactedOnPurpose"
      },
      "deploymentId" : RedactedOnPurpose
    }
  ],
  "experiments" : [
    {
      "treatmentId" : "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxx2abd",
      "experimentId" : "RedactedOnPurpose",
      "deploymentId" : RedactedOnPurpose
    },
    {
      "treatmentId" : "d3835ae6-d7dd-4c39-8e74-f3fead549120",
      "experimentId" : "631fac54de559130376df80f",
      "deploymentId" : 400000038
    }
  ]
}
}

All of my devices, iPads and iPhone are running a non standard OS. Apps downloaded from the AppStore have been swapped out for beta version, evidenced by Beta Identifier numbers, and experiments being pushed to it. Even SpringBoard is a beta version. Unfortunately, it has been able to migrate to family and friends devices who were in my contacts. To date, nothing has been able to remove them. 

I suspect the exploit source is Apples proprietary ’triald’ process. Not much is publicly known about what it exists for, however it appears it has capabilities to modify the OS as a developer. See plist for siritts and triald plists.